use class without dependencies
